+++ app/src/main/java/com/seravian/seravianapp/SeravianApp.kt
@@ -3,14 +3,15 @@ package com.seravian.seravianapp
 import android.app.Application
 import com.greenvenom.core_network.api.di.apiModule
 import com.seravian.seravianapp.di.appModule
-import com.greenvenom.feat_auth.di.authFeatureModule
+import com.greenvenom.feat_auth.di.authModule
 import com.greenvenom.core_navigation.di.navigationCoreModule
 import com.seravian.feat_home.di.homeModule
-import com.greenvenom.feat_onboarding.di.onBoardingFeatureModule
+import com.greenvenom.feat_onboarding.di.onBoardingModule
+import com.greenvenom.feat_tokens.di.tokensModule
 import com.seravian.feat_chat.di.chatModule
-import com.seravian.feat_local.di.localFeatureModule
+import com.seravian.feat_local.di.localModule
 import com.seravian.feat_navigation.di.navigationFeatureModule
-import com.seravian.feat_network.di.networkFeatureModule
+import com.seravian.feat_network.di.networkModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
@@ -25,12 +26,13 @@ class SeravianApp: Application() {
             modules(
                 appModule,
                 apiModule,
-                localFeatureModule,
-                networkFeatureModule,
+                tokensModule,
+                localModule,
+                networkModule,
                 navigationCoreModule,
                 navigationFeatureModule,
-                authFeatureModule,
-                onBoardingFeatureModule,
+                authModule,
+                onBoardingModule,
                 homeModule,
                 chatModule
             )
