+++ auth/feat-auth/src/main/java/com/greenvenom/feat_auth/di/AuthModule.kt
@@ -1,6 +1,8 @@
 package com.greenvenom.feat_auth.di
 
-import com.greenvenom.core_auth.data.repository.EmailStateRepository
+import com.greenvenom.feat_auth.data.repository.EmailStateRepository
+import com.greenvenom.feat_auth.domain.repository.AuthRepository
+import com.greenvenom.feat_auth.data.repository.AuthRepositoryImpl
 import com.greenvenom.feat_auth.presentation.login.LoginViewModel
 import com.greenvenom.feat_auth.presentation.otp.OtpViewModel
 import com.greenvenom.feat_auth.presentation.register.RegisterViewModel
@@ -8,8 +10,17 @@ import com.greenvenom.feat_auth.presentation.reset_password.ResetPasswordViewMod
 import org.koin.core.module.dsl.viewModel
 import org.koin.dsl.module
 
-val authFeatureModule = module {
+val authModule = module {
     single { EmailStateRepository() }
+
+    single<AuthRepository> {
+        AuthRepositoryImpl(
+            remoteDataSource = get(),
+            roomDataSource = get(),
+            tokensRepository = get(),
+            emailStateRepository = get()
+        )
+    }
 
     viewModel { LoginViewModel(authRepository = get(), emailStateRepository = get()) }
     viewModel { RegisterViewModel(emailStateRepository = get(), authRepository = get()) }
