+++ auth/feat-auth/src/main/java/com/greenvenom/feat_auth/presentation/login/LoginViewModel.kt
@@ -2,8 +2,8 @@ package com.greenvenom.feat_auth.presentation.login
 
 import androidx.lifecycle.viewModelScope
 import com.greenvenom.core_auth.data.dto.request.LoginRequest
-import com.greenvenom.core_auth.data.repository.EmailStateRepository
-import com.greenvenom.core_auth.domain.repository.AuthRepository
+import com.greenvenom.feat_auth.data.repository.EmailStateRepository
+import com.greenvenom.feat_auth.domain.repository.AuthRepository
 import com.greenvenom.core_ui.presentation.BaseViewModel
 import com.greenvenom.validation.ValidateInput
 import kotlinx.coroutines.flow.MutableStateFlow
