+++ network/core-network/build.gradle.kts
@@ -61,6 +61,7 @@ dependencies {
     implementation(project(":onboarding:core-onboarding"))
 
     implementation(libs.androidx.core.ktx)
+
     implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.8.0")
 
     implementation("io.ktor:ktor-client-cio:$ktor")
