+++ onboarding/feat-onboarding/build.gradle.kts
@@ -35,11 +35,12 @@ android {
 
 dependencies {
 
+    implementation(project(":onboarding:core-onboarding"))
     implementation(project(":network:core-network"))
     implementation(project(":local:core-local"))
     implementation(project(":tokens:core-tokens"))
+    implementation(project(":profile:core-profile"))
     implementation(project(":validation"))
-    implementation(project(":onboarding:core-onboarding"))
     implementation(project(":core-ui"))
 
     implementation(libs.androidx.core.ktx)
