--- profile/core-profile/src/main/java/com/seravian/core_profile/di/ProfileModule.kt
@@ -1,7 +1,0 @@
-package com.seravian.core_profile.di
-
-import org.koin.dsl.module
-
-val profileModule = module {
-
-}
