+++ tokens/feat-tokens/build.gradle.kts
@@ -36,8 +36,8 @@ android {
 dependencies {
     val koin = "4.0.0"
 
-    implementation(project(":network:core-network"))
     implementation(project(":tokens:core-tokens"))
+    implementation(project(":network:core-network"))
     implementation(project(":crypto"))
 
     implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.8.0")
